FEAT -- rename name of the files into some components
