+++ src/components/Button/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import clsx from "clsx";
-import { ButtonProps } from "@/models/attributes-props";
+import { ButtonProps } from "@/models/props";
 import stylesheet from "./stylesheet.module.scss";
 
 export const Button: React.FC<ButtonProps> = ({
