+++ src/components/Container/index.tsx
@@ -1,15 +1,20 @@
 import React from "react";
 import clsx from "clsx";
-import { ContainerProps } from "@/models/attributes-props";
+import { ContainerProps } from "@/models/props";
 import stylesheet from "./stylesheet.module.scss";
 
-export const Container: React.FC<ContainerProps> = ({
+const Container: React.FC<ContainerProps> = ({
+  row,
+  column,
   className,
   children,
   ...props
 }) => {
   return (
-    <div className={clsx(stylesheet["container"], className)} {...props}>
+    <div className={clsx(stylesheet["container"], className, {
+      [stylesheet['row']]: row,
+      [stylesheet['column']]: column
+    })} {...props}>
       {children}
     </div>
   );
