+++ src/components/Image/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import clsx from "clsx";
-import { ImageProps } from "@/models/attributes-props";
+import { ImageProps } from "@/models/props";
 import stylesheet from "./stylesheet.module.scss";
 
 export const Image: React.FC<ImageProps> = ({
