+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { Routes } from "@/helpers/system/routes";
-import reportWebVitals from "./reportWebVitals";
+import Routes from "#/src/routes";
+import reportWebVitals from "#/src/reportWebVitals";
 import "@/assets/stylesheets/global.scss";
 
 ReactDOM.render(
