+++ src/pages/SignIn/index.tsx
@@ -3,9 +3,10 @@ import clsx from "clsx";
 import stylesheet from "./stylesheet.module.scss";
 import { DefaultState } from "@/models/default";
 import Loader from "@/components/Loader";
+import Container from "@/components/Container";
 import Image from "@/components/Image";
 import LogoImage from "@/assets/images/logo.jpg";
-import Title from "#/src/components/Title";
+import Title from "@/components/Title";
 
 interface State extends DefaultState {
   title: string;
@@ -27,14 +28,14 @@ export const SignIn: React.FC = () => {
   return state.pending === true ? (
     <Loader />
   ) : (
-    <>
+    <Container>
       <Title name={state.title} />
       <Image
         image={LogoImage}
         alt="logo"
         className={clsx(stylesheet["logo"])}
       />
-    </>
+    </Container>
   );
 };
